+++ packages/ai/src/ui/index.ts
@@ -31,8 +31,8 @@ export { TextStreamChatTransport } from './text-stream-chat-transport';
 export {
   getToolName,
   getToolOrDynamicToolName,
-  isToolUIPart,
   isToolOrDynamicToolUIPart,
+  isToolUIPart,
   type DataUIPart,
   type DynamicToolUIPart,
   type FileUIPart,
@@ -48,6 +48,7 @@ export {
   type UIMessage,
   type UIMessagePart,
   type UITool,
+  type UIToolInvocation,
   type UITools,
 } from './ui-messages';
 export {