+++ packages/openai/src/responses/openai-responses-api-types.ts
@@ -13,6 +13,19 @@ export type OpenAIResponsesMessage =
   | OpenAIFileSearchCall
   | OpenAIResponsesReasoning;
 
+export type OpenAIResponsesIncludeOptions =
+  | Array<
+      | 'web_search_call.action.sources'
+      | 'code_interpreter_call.outputs'
+      | 'computer_call_output.output.image_url'
+      | 'file_search_call.results'
+      | 'message.input_image.image_url'
+      | 'message.output_text.logprobs'
+      | 'reasoning.encrypted_content'
+    >
+  | undefined
+  | null;
+
 export type OpenAIResponsesSystemMessage = {
   role: 'system' | 'developer';
   content: string;
@@ -80,6 +93,22 @@ export type OpenAIResponsesTool =
       description: string | undefined;
       parameters: JSONSchema7;
       strict?: boolean;
+    }
+  | {
+      type: 'web_search';
+      filters?: {
+        allowed_domains?: string[];
+      };
+      search_context_size: 'low' | 'medium' | 'high' | undefined;
+      user_location:
+        | {
+            type: 'approximate';
+            city?: string;
+            country?: string;
+            region?: string;
+            timezone?: string;
+          }
+        | undefined;
     }
   | {
       type: 'web_search_preview';